refactor: split `scaledZeroOrMinOrMax` from `zeroOrMinOrMaxPosition` to avoid unreachable lines
